Fix threading bug in DefaultInferenceParameters#getParams
DefaultInferenceParameters#getParams would always return the same cached instance of DefaultInferenceParameters. Thus, if two queries were running in different threads, each would acquire the same instance of DefaultInferenceParameters, and if one thread modified it, it would cause a ConcurrentModificationException in object's the #hashcode method. More generally, it is problematic if, when you modify one instance of DefaultInferenceParameters, you modify *all existing instances* of DefaultInferenceParameters.

This patch should fix that. It does two things:

- paramCache is now a ConcurrentHashMap
- getParams now returns a #cloned copy of the cached CycList (which is used to construct the object) rather than the actual instance itself.

In other words: the params are cached, but each instance of DefaultInferenceParameters receives its own independent copy of them.